dev/requirements.txt
cassandra_driver==3.25.0
Django==3.2.5
django_pandas==0.6.6
pandas==1.3.1
django-crispy-forms==1.12.0
psycopg2==2.9.1
django-cassandra-engine==1.7.0
XlsxWriter==3.0.3
openpyxl==3.0.10